add to Type Of Work Api delete and update services

+++ src/main/java/com/umutdoruk/hrms/repository/TypeOfWorkRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypeOfWorkRepository extends JpaRepository<TypeOfWork,Integer> {
+public interface TypeOfWorkRepository extends JpaRepository<TypeOfWork,Long> {
 }
